refactoring

+++ src/main/java/spimedb/util/math/MathUtils.java
@@ -519,11 +519,11 @@ public final class MathUtils {
     }
 
     public static int random(Random rnd, int max) {
-        return rnd.nextInt((int) max);
+        return rnd.nextInt(max);
     }
 
     public static int random(Random rnd, int min, int max) {
-        return rnd.nextInt((int) (max - min)) + min;
+        return rnd.nextInt(max - min) + min;
     }
 
     public static boolean randomChance(double chance) {
